+++ ffi/src/sspi/macros.rs
