+++ ffi/src/sspi/sec_buffer.rs
