+++ ffi/src/sspi/sec_pkg_info.rs
@@ -6,8 +6,8 @@ use sspi::{enumerate_security_packages, PackageInfo, KERBEROS_VERSION};
 #[cfg(windows)]
 use symbol_rename_macro::rename_symbol;
 
-use crate::sspi_data_types::{SecChar, SecWChar, SecurityStatus};
-use crate::utils::c_w_str_to_string;
+use super::sspi_data_types::{SecChar, SecWChar, SecurityStatus};
+use super::utils::c_w_str_to_string;
 
 #[derive(Debug)]
 #[repr(C)]
