+++ ffi/src/sspi/sspi_data_types.rs
